requirements.txt
APScheduler==3.8.1
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.10.8
click==8.0.3
colorama==0.4.4
colorhash==1.0.3
config==0.5.1
configparser==5.1.0
dnspython==2.1.0
Flask==2.0.2
Flask-Cors==3.0.10
Flask-MonitoringDashboard==3.1.1
greenlet==1.1.2
importlib-metadata==4.8.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.21.4
packaging==21.3
pandas==1.3.4
pluggy==1.0.0
psutil==5.8.0
py==1.11.0
pymongo==3.12.1
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
scipy==1.7.2
secure-smtplib==0.1.1
six==1.16.0
SQLAlchemy==1.4.27
toml==0.10.2
typing_extensions==4.0.0
tzlocal==2.0.0
Werkzeug==2.0.2
wincertstore==0.2
zipp==3.6.0
